+++ src/main/java/com/cpg/pprojects/ecommerce/entity/cart/repository/ICartRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ICartRepository {
     List<Cart> findAll();
-    Cart findById(int id);
+    Cart findById(long id);
     Cart save(Cart cart);
     void delete(Cart cart);
